Text To Speech/requirements.txt
fastapi
uvicorn
python-multipart
azure-cognitiveservices-speech